Updated the config provider
